chore: update dependencies to latest major versions
- reqwest 0.12 -> 0.13 (feature rustls-tls renamed to rustls)
- rusqlite 0.32 -> 0.40 (u64 to i64 casts for ToSql compat)
- sha2 0.10 -> 0.11
- vite 6.x -> 8.x (security: 2 high CVEs fixed)
- tailwindcss 4.3.0 -> 4.3.1
- Various transitive updates via cargo update

+++ src/governance/quotas.rs
@@ -247,7 +247,7 @@ impl Database {
                  tokens_per_hour = excluded.tokens_per_hour,
                  tokens_per_day = excluded.tokens_per_day,
                  tokens_per_month = excluded.tokens_per_month",
-            params![id, user_id, tokens_per_hour, tokens_per_day, tokens_per_month],
+            params![id, user_id, tokens_per_hour as i64, tokens_per_day as i64, tokens_per_month as i64],
         )?;
 
         Ok(())
